+++ internal/transaction/injector/wire.go
@@ -8,6 +8,7 @@ import (
 	cart "Dzaakk/simple-commerce/internal/shopping_cart/repository"
 	handler "Dzaakk/simple-commerce/internal/transaction/handler"
 	repository "Dzaakk/simple-commerce/internal/transaction/repository"
+	routes "Dzaakk/simple-commerce/internal/transaction/routes"
 	usecase "Dzaakk/simple-commerce/internal/transaction/usecase"
 	"database/sql"
 
@@ -22,6 +23,7 @@ func InitializedService(db *sql.DB) *handler.TransactionHandler {
 		cart.NewShoppingCartRepository,
 		cart.NewShoppingCartItemRepository,
 		customer.NewCustomerRepository,
+		routes.NewTransactionRoutes,
 	)
 
 	return &handler.TransactionHandler{}
